rename ActivateUdbOrganizer -> ApproveUdbOrganizer
